tests/wpt/web-platform-tests/fetch/sec-metadata/window-open.tentative.https.sub.html

<!DOCTYPE html>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/resources/testdriver.js></script>
<script src=/resources/testdriver-vendor.js></script>
<script src=/fetch/sec-metadata/resources/helper.js></script>
<body>
<script>
  // Forced navigations:
  async_test(t => {
    let w = window.open("https://{{host}}:{{ports[https][0]}}/fetch/sec-metadata/resources/post-to-owner.py");
    t.add_cleanup(_ => w.close());
    window.addEventListener('message', t.step_func(e => {
      if (e.source != w)
        return;

      assert_header_equals(e.data, {
        "dest": "document",
        "site": "same-origin",
        "user": "",
        "mode": "navigate",
      });
      t.done();
    }));
  }, "Same-origin window, forced");

  async_test(t => {
    let w = window.open("https://{{hosts[][www]}}:{{ports[https][0]}}/fetch/sec-metadata/resources/post-to-owner.py");
    t.add_cleanup(_ => w.close());
    window.addEventListener('message', t.step_func(e => {
      if (e.source != w)
        return;

      assert_header_equals(e.data, {
        "dest": "document",
        "site": "same-site",
        "user": "",
        "mode": "navigate",
      });
      t.done();
    }));
  }, "Same-site window, forced");

  async_test(t => {
    let w = window.open("https://{{hosts[alt][www]}}:{{ports[https][0]}}/fetch/sec-metadata/resources/post-to-owner.py");
    t.add_cleanup(_ => w.close());
    window.addEventListener('message', t.step_func(e => {
      if (e.source != w)
        return;

      assert_header_equals(e.data, {
        "dest": "document",
        "site": "cross-site",
        "user": "",
        "mode": "navigate",
      });
      t.done();
    }));
  }, "Cross-site window, forced");

  async_test(t => {
    let w = window.open("https://{{host}}:{{ports[https][0]}}/fetch/sec-metadata/resources/post-to-owner.py");
    t.add_cleanup(_ => w.close());
    let messages = 0;
    window.addEventListener('message', t.step_func(e => {
      messages++;
      if (e.source != w)
        return;

      assert_header_equals(e.data, {
        "dest": "document",
        "site": "same-origin",
        "user": "",
        "mode": "navigate",
      });

      if (messages == 1) {
        w.location.reload();
      } else {
        t.done();
      }
    }));
  }, "Same-origin window, forced, reloaded");

  async_test(t => {
    let w = window.open("https://{{hosts[][www]}}:{{ports[https][0]}}/fetch/sec-metadata/resources/post-to-owner.py");
    t.add_cleanup(_ => w.close());
    let messages = 0;
    window.addEventListener('message', t.step_func(e => {
      messages++;
      if (e.source != w)
        return;

      assert_header_equals(e.data, {
        "dest": "document",
        "site": "same-site",
        "user": "",
        "mode": "navigate",
      });

      if (messages == 1) {
        w.location.reload();
      } else {
        t.done();
      }
    }));
  }, "Same-site window, forced, reloaded");

  async_test(t => {
    let w = window.open("https://{{hosts[alt][www]}}:{{ports[https][0]}}/fetch/sec-metadata/resources/post-to-owner.py");
    t.add_cleanup(_ => w.close());
    let messages = 0;
    window.addEventListener('message', t.step_func(e => {
      messages++;
      if (e.source != w)
        return;

      assert_header_equals(e.data, {
        "dest": "document",
        "site": "cross-site",
        "user": "",
        "mode": "navigate",
      });

      if (messages == 1) {
        w.location.reload();
      } else {
        t.done();
      }
    }));
  }, "Cross-site window, forced, reloaded");

  // User-activated navigations:
  async_test(t => {
    let b = document.createElement('button');
    b.onclick = t.step_func(_ => {
      let w = window.open("https://{{host}}:{{ports[https][0]}}/fetch/sec-metadata/resources/post-to-owner.py");
      t.add_cleanup(_ => w.close());
      window.addEventListener('message', t.step_func(e => {
        if (e.source != w)
          return;

        assert_header_equals(e.data, {
          "dest": "document",
          "site": "same-origin",
          "user": "?T",
          "mode": "navigate",
        });
        t.done();
      }));
    });
    document.body.appendChild(b);
    test_driver.click(b);
  }, "Same-origin window, user-activated");

  async_test(t => {
    let b = document.createElement('button');
    b.onclick = t.step_func(_ => {
      let w = window.open("https://{{hosts[][www]}}:{{ports[https][0]}}/fetch/sec-metadata/resources/post-to-owner.py");
      t.add_cleanup(_ => w.close());
      window.addEventListener('message', t.step_func(e => {
        if (e.source != w)
          return;

        assert_header_equals(e.data, {
          "dest": "document",
          "site": "same-site",
          "user": "?T",
          "mode": "navigate",
        });
        t.done();
      }));
    });
    document.body.appendChild(b);
    test_driver.click(b);
  }, "Same-site window, user-activated");

  async_test(t => {
    let b = document.createElement('button');
    b.onclick = t.step_func(_ => {
      let w = window.open("https://{{hosts[alt][www]}}:{{ports[https][0]}}/fetch/sec-metadata/resources/post-to-owner.py");
      t.add_cleanup(_ => w.close());
      window.addEventListener('message', t.step_func(e => {
        if (e.source != w)
          return;

        assert_header_equals(e.data, {
          "dest": "document",
          "site": "cross-site",
          "user": "?T",
          "mode": "navigate",
        });
        t.done();
      }));
    });
    document.body.appendChild(b);
    test_driver.click(b);
  }, "Cross-site window, user-activated");
</script>
